Attempt to insert second teleportal into pair
There is a bug in the Sheetsu API that prevents the search function from working as advertised. In particular, When there is just one row and it contains the target field then the search works. But when there is more than one field and one of them does not match then the search returns nothing.